Type conversion, part 2

+++ core/mem.c
@@ -24,7 +24,7 @@ static int mem_io(sl_map_ep_t *ep, sl_io_op_t *op) {
     return sl_io_for_data(data, op);
 }
 
-int mem_region_create(u64 base, u64 length, mem_region_t **m_out) {
+int mem_region_create(u8 base, u8 length, mem_region_t **m_out) {
     const size_t size = length + sizeof(mem_region_t);
     mem_region_t *m = calloc(1, size);
     if (m == NULL) return SL_ERR_MEM;
